\name{hapmap_geno}
\docType{data}
\alias{hapmap_geno}
\title{SNP genotypes of HapMap samples}
\description{
    A list object including the following components:

    \code{sample.id} -- a vector of sample ids;

    \code{snp.id} -- a vector of SNP ids;

    \code{snp.position} -- a vector of SNP positions;

    \code{snp.chromosome} -- a vector of chromosome indices;

    \code{snp.allele} -- a character vector of ``reference / non-reference'';

    \code{genotype} -- a ``# of SNPs'' X ``# of samples'' genotype matrix.
}
\usage{hapmap_geno}
\value{A list}
\keyword{datasets}
