\name{snpgdsExampleFileName}
\alias{snpgdsExampleFileName}
\title{Example GDS file}
\description{
	Return the file name of example data
}
\usage{
snpgdsExampleFileName()
}
\details{
	A GDS genotype file is created from a subset of HapMap Phase II dataset consisting of
270 individuals and duplicates.
}
\author{Xiuwen Zheng \email{zhengx@u.washington.edu}}

\keyword{gds}
\keyword{GWAS}
