\name{snpgdsPosition}
\alias{snpgdsPosition}
\title{
	SNP positions in ascending order
}
\description{
	Make SNP positions ascending.
}
\usage{
snpgdsPosition(gds)
}
\arguments{
	\item{gds}{the \code{\link[gdsfmt]{gdsclass}} object in the \link{gdsfmt} package, or a file name}
}
\value{
	None.
}
\author{Xiuwen Zheng \email{zhengx@u.washington.edu}}


\keyword{gds}
\keyword{GWAS}
