\name{snpgdsExampleFileName}
\alias{snpgdsExampleFileName}
\title{Example GDS file}
\description{
    Return the file name of example data
}
\usage{
snpgdsExampleFileName()
}
\details{
    A GDS genotype file was created from a subset of HapMap Phase II dataset
consisting of 270 individuals and duplicates.
}
\value{Characters}

\author{Xiuwen Zheng}

\examples{
snpgdsExampleFileName()
}

\keyword{GDS}
\keyword{GWAS}
